increase stack (for x86-64 ELF), add simple prefix sum

+++ a+b_device_program.cl
@@ -1,3 +1,3 @@
-kernel void add(global const int *a, global const int *b, global int *c) {
+kernel void add(global const float8 *a, global const float8 *b, global float8 *c) {
     *c = *a + *b;
 }
